# doc/openapi.yaml
swagger: "2.0"
info:
  version: 9.7.0
  title: "@fastify/swagger"
...
paths:
  /graphql:
    get:
      parameters:
        - type: string
          required: false
          in: query
          name: query
        - type: boolean
          required: false
          in: query
          name: persisted
        - type:
            - string
            - "null"
          required: false
          in: query
          name: operationName
        - type: string
          required: false
          in: query
          name: variables
        - type: string
          required: false
          in: query
          name: extensions
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              data:
                type:
                  - "null"
                  - object
                additionalProperties: true
              errors:
                type: array
                items:
                  type: object
                  required:
                    - message
                  properties:
                    message:
                      type: string
                    locations:
                      type: array
                      items:
                        type: object
                        properties:
                          line:
                            type: integer
                          column:
                            type: integer
                    path:
                      type: array
                      items:
                        type: string
                    extensions:
                      type: object
                      properties:
                        code:
                          type: string
                      additionalProperties: true
              extensions:
                type: object
                additionalProperties: true
    post:
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              query:
                type: string
              persisted:
                type: boolean
              operationName:
                type:
                  - string
                  - "null"
              variables:
                type:
                  - object
                  - "null"
              extensions:
                type: object
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            properties:
              data:
                type:
                  - "null"
                  - object
                additionalProperties: true
              errors:
                type: array
                items:
                  type: object
                  required:
                    - message
                  properties:
                    message:
                      type: string
                    locations:
                      type: array
                      items:
                        type: object
                        properties:
                          line:
                            type: integer
                          column:
                            type: integer
                    path:
                      type: array
                      items:
                        type: string
                    extensions:
                      type: object
                      properties:
                        code:
                          type: string
                      additionalProperties: true
              extensions:
                type: object
                additionalProperties: true
...
  /api/supervisors/l402-readiness/:
    get:
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            required:
              - domainId
              - architecture
              - webhookEndpoint
              - secretManagerPath
              - checklistApprovals
              - diagnostics
            properties:
              id:
                type: number
              domainId:
                type: number
              architecture:
                type: string
              webhookEndpoint:
                anyOf:
                  - type: string
                  - type: "null"
              secretManagerPath:
                anyOf:
                  - type: string
                  - type: "null"
              checklistApprovals:
                type: object
                additionalProperties:
                  type: string
              diagnostics:
                type: object
                required:
                  - hasLnbitsUrl
                  - hasLnbitsKey
                  - webhookConfigured
                properties:
                  hasLnbitsUrl:
                    type: boolean
                  hasLnbitsKey:
                    type: boolean
                  webhookConfigured:
                    type: boolean
        "400":
          description: Default Response
          schema:
            type: object
            required:
              - error
              - code
              - remediation
            properties:
              error:
                type: string
              code:
                type: string
              remediation:
                type: string
        "401":
          description: Default Response
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    put:
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              architecture:
                type: string
                enum:
                  - mock
                  - lnbits
              webhookEndpoint:
                type: string
              secretManagerPath:
                type: string
              checklistApprovals:
                type: object
                additionalProperties:
                  type: string
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            required:
              - ok
            properties:
              ok:
                type: boolean
                enum:
                  - true
        "400":
          description: Default Response
          schema:
            type: object
            required:
              - error
              - code
              - remediation
            properties:
              error:
                type: string
              code:
                type: string
              remediation:
                type: string
        "401":
          description: Default Response
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string